[REFACTOR] #48 - 사이드 / 국물요리 화면 isCartAdded 체크 로직 및 combine ViewModel 내에 작업
