+++ app/src/main/java/com/example/banchan/presentation/home/soup/SoupViewModel.kt
@@ -3,19 +3,23 @@ package com.example.banchan.presentation.home.soup
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.banchan.R
-import com.example.banchan.domain.usecase.GetSoupDishesUseCase
+import com.example.banchan.data.source.local.basket.BasketItem
+import com.example.banchan.domain.usecase.basket.GetBasketItemUseCase
+import com.example.banchan.domain.usecase.home.GetSoupDishesUseCase
 import com.example.banchan.presentation.adapter.home.CommonItemListModel
 import com.example.banchan.presentation.home.maincook.Filter
 import com.example.banchan.util.ext.toNum
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 class SoupViewModel @Inject constructor(
-    private val getSoupDishesUseCase: GetSoupDishesUseCase
+    private val getSoupDishesUseCase: GetSoupDishesUseCase,
+    private val getBasketItemUseCase: GetBasketItemUseCase
 ) : ViewModel() {
     private var filter = Filter.NORMAL
     private val _dishes =
@@ -26,6 +30,14 @@ class SoupViewModel @Inject constructor(
             )
         )
     val dishes = _dishes.asStateFlow()
+
+    val commonListItems =
+        combine(dishes, getBasketItemUseCase()) { dishes, basketList ->
+            basketList.onSuccess {
+                return@combine checkIsCartAdded(dishes, it)
+            }
+            dishes
+        }
 
     init {
         viewModelScope.launch {
@@ -63,4 +75,21 @@ class SoupViewModel @Inject constructor(
             CommonItemListModel.SmallItem(it)
         }
     }
+
+    private fun checkIsCartAdded(
+        dishes: List<CommonItemListModel>,
+        basketList: List<BasketItem>
+    ): List<CommonItemListModel> =
+        dishes.map { dish ->
+            if (dish is CommonItemListModel.SmallItem) {
+                dish.copy(
+                    item = dish.item.copy(
+                        isCartAdded = dish.item.detailHash in basketList.map { it.hash }
+                    )
+                )
+            } else {
+                dish
+            }
+        }
+
 }
